Milestone 1
### 1. Register and Load Users:
Includes a registration system that securely stores user data, including salted and hashed passwords, using the Crypto++ library. The Boost's serialization library is used for efficient data storage and retrieval. A login feature authenticates users based on the stored data, performing the inverse of the data storage process to compare.